+++ result.h
@@ -33,7 +33,7 @@ struct result_t {
 struct result_t *result_new(char *name, size_t initial_size);
 int result_add(struct result_t *result, char *name);
 void result_free(struct result_t *result);
-int result_print(struct result_t *result);
+size_t result_print(struct result_t *result);
 int result_cmp(const void *r1, const void *r2);
 
 #endif /* _NOSR_RESULT_H */
